Upgrade jTravis version (#541)
* Move every Travis ID from int to long

* Should fix failign tests

+++ repairnator/repairnator-dockerpool/src/main/java/fr/inria/spirals/repairnator/dockerpool/AbstractPoolManager.java
@@ -83,7 +83,7 @@ public class AbstractPoolManager {
         this.engines = engines;
     }
 
-    public TreatedBuildTracking prepareBeforeSubmitBuild(int buildId) {
+    public TreatedBuildTracking prepareBeforeSubmitBuild(long buildId) {
         this.cleanUpOlderContainers();
         return new TreatedBuildTracking(this.engines, this.runId, buildId);
     }
