+++ repairnator/repairnator-pipeline/src/test/java/fr/inria/spirals/repairnator/process/step/paths/TestComputeClasspath.java
@@ -50,7 +50,7 @@ public class TestComputeClasspath {
 
     @Test
     public void testComputeClasspath() throws IOException {
-        int buildId = 201176013; // surli/failingProject build
+        long buildId = 201176013; // surli/failingProject build
 
         Optional<Build> optionalBuild = RepairnatorConfig.getInstance().getJTravis().build().fromId(buildId);
         assertTrue(optionalBuild.isPresent());
